Updated working with session

+++ server/api/controllers/questions/controller.ts
@@ -4,13 +4,16 @@ import { Request, Response } from 'express';
 export class Controller {
   all(req: Request, res: Response): void {
     DbService.getAllQuestions().then(r => {
-      r['email'] = req['session'] && req['session'].email;
+      r['given_name'] = req['session'] && req['session'].given_name;
       return res.json(r);
     });
   }
 
-  create(req: Request, res: Response): void {
-    DbService.createQuestion(req.body).then(r =>
+  create(req: Request, res: Response) {
+    if (!req['session'] || !req['session'].email) {
+      return res.status(401).json({ error: "401 Unauthorized" });
+    }
+    DbService.createQuestion(req.body, req['session']).then(r =>
       res
         .status(201)
         .location(`/api/v1/questions/${r}`)
@@ -18,8 +21,11 @@ export class Controller {
     );
   }
 
-  update(req: Request, res: Response): void {
-    DbService.updateQuestion(req.body).then(r =>
+  update(req: Request, res: Response) {
+    if (!req['session'] || !req['session'].email) {
+      return res.status(401).json({ error: "401 Unauthorized" });
+    }
+    DbService.updateQuestion(req.body, req['session']).then(r =>
       res
         .status(201)
         .location(`/api/v1/questions/${r}`)
@@ -27,7 +33,10 @@ export class Controller {
     );
   }
 
-  delete(req: Request, res: Response): void {
+  delete(req: Request, res: Response) {
+    if (!req['session'] || !req['session'].email) {
+      return res.status(401).json({ error: "401 Unauthorized" });
+    }
     DbService.deleteQuestion(req.body.id).then(r =>
       res
         .status(201)
